Reverted most of the crammed 15-bit lfsr_attr_t representation
I think this was becoming too over-engineered. That and if you're
fighting the compiler too much, sometimes it's best to just let the
compiler win.

There is also the minor benefit of not needing to worry about uint16_t
overflows in lfsr_cat_t. One concern is it's easy for missed lfsr_data_t
indirections to go unnoticed.

This reverts lfsr_attr_t from 3 words -> 4 words, but allows easy
mapping to the lfsr_cat_t part of the lfsr_attr_t:

  lfsr_tag_t               lfsr_attr_t
  .---+---.                .---+---+---+---. . . .---+---+---+---.
  |  tag  |--------------->|  tag  |       |     |  tag  |       |
  '---+---'                +---+---+---+---+     +---+---+---+---+
                    .----->|     delta     |     |     delta     |
  lfsr_srid_t       |      +---+---+---+---+     +---+---+---+---+
  .---+---+---+---. | .--->|c|    size     |     |      cat      |
  |     delta     |-' |    +---+---+---+---+     |               |
  '---+---+---+---'   | .->|      ptr      |     |               |
                      | |  '---+---+---+---' . . '---+---+---+---'
  lfsr_cat_t          | |
  .---+---+---+---.   | |
  |c|    size     |---' |
  +---+---+---+---+     |
  |      ptr      |-----'
  '---+---+---+---'

This doesn't impact stack as much as you would expect due to compiler
overhead around inline functions operating on multi-word structs:

  before 15-bit: 33672           2776
  best 15-bit:   33900 (+0.7%)   2736 (-1.4%)
  prev 15-bit:   34486 (+2.4%)   2912 (+4.9%)
  after:         33672 (+0.0%)   2744 (-1.2%)

I think the only big thing not reverted was switching lfsr_cat_t from
storing the sum of the lfsr_data_ts sizes to just storing the number of
lfsr_data_ts. This is cheaper to initialize (because inline functions
are weirdly costly!), cheaper to iterate over, and we don't actually
need lfsr_cat_size that often, unlike lfsr_data_size.

Eventually we'll probably need to revisit this to optimize 31/15-bit,
etc, littlefs configurations. But that's a problem for another day...