Tweaked lfsr_format to use 0x00216968 for initial revision count
The main reason is just to avoid using 0x00000000 as the initial
revision count. The is currently the default for B-tree rbyds, and
accidentally writing a B-tree rbyd to an mroot block is both easy
(misconfigured block_count) and something we really want to notice when
debugging.

Fortunately we can still keep our sequence comparison test (0 > -1) by
only setting the top-bits. Though we still need to zero the recycle
counter to avoid premature mroot extension, so this magic number may not
stay intact depending on configuration.

This does add a bit of code, probably to load the magic number from
Thumb's constant pools, but I think it's worth it:

           code          stack
  before: 33430           2640
  after:  33458 (+0.1%)   2640 (+0.0%)