Allowed mdir overcompaction if we run out of space
This should allow mdir commits that would normally trigger a relocation
to continue if lfsr_mdir_alloc__ return LFS_ERR_NOSPC, though at least
with a logged warning.

This seems preferable to the alternative: locking up the filesystem.

Though this doesn't have tests yet, so take it with a grain of salt...

Code changes minimal:

           code          stack
  before: 33470           2640
  after:  33474 (+0.0%)   2640 (+0.0%)