Lifted block range checks up into rbyd append functions
The only real use case for the bd runtime bounds checks is to abort rbyd
commits when they run off the end of the block. Since rbyd's now have
their own set of low-level append functions, we're better off doing the
bounds checks there and changing all of the lfsr_bd_* bounds checks to
asserts.

Block overflows are a particularly easy mistake to make, and one that
would be good to catch early.

One interesting thing to note: We're now using LFSR_TAG_DSIZE for range
checks instead of the actual tag encoding. This may seem suboptimal, but
if LFSR_TAG_DSIZE can't fit in the remaining space in the block, the
cksum tag wouldn't be able to fit anyways. So we're not really wasting
any space.

This saves a nice bit of code:

           code          stack
  before: 33690           2608
  after:  33610 (-0.2%)   2608 (+0.0%)